+++ packages/eyes-sdk-core/lib/utils/createFramesPaths.js
@@ -1,6 +1,6 @@
 function createFramesPaths({snapshot, path = [], logger}) {
-  const paths = snapshot.crossFramesXPaths
-    ? snapshot.crossFramesXPaths.map(selector => ({
+  const paths = snapshot.crossFramesSelectors
+    ? snapshot.crossFramesSelectors.map(selector => ({
         path: path.concat(selector),
         parentSnapshot: snapshot,
       }))
@@ -13,7 +13,7 @@ function createFramesPaths({snapshot, path = [], logger}) {
   }
 
   logger.verbose(
-    `frames paths for ${snapshot.crossFramesXPaths}`,
+    `frames paths for ${snapshot.crossFramesSelectors}`,
     paths.map(({path}) => path.join('-->')).join(' , '),
   )